# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bedrock SSO Proxy API",
    "description": "JWT-authenticated HTTP proxy server that provides secure access to AWS Bedrock APIs",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/usage/records": {
      "get": {
        "tags": [
          "Admin Usage Management"
        ],
        "summary": "Get System Usage Records",
        "description": "Retrieve usage records for all users (admin only)",
        "operationId": "get_system_usage_records",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to return (default: 50, max: 500)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of records to skip for pagination",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "model",
            "in": "query",
            "description": "Filter by specific model ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "description": "Filter records from this date onwards",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "Filter records up to this date",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "success_only",
            "in": "query",
            "description": "If true, only return successful requests",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Response format (json or csv)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "Filter by specific user ID (admin only)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of system usage records",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageRecordsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_auth": []
          }
        ]
      }
    },
    "/api/admin/usage/summaries": {
      "get": {
        "tags": [
          "Admin Usage Management"
        ],
        "summary": "Get System Usage Summaries",
        "description": "Retrieve pre-computed usage summaries for all users (admin only)",
        "operationId": "get_admin_usage_summaries",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "description": "Filter summaries from this date onwards",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "Filter summaries up to this date",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "period_type",
            "in": "query",
            "description": "Period type for aggregation (hourly, daily, weekly, monthly)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "model_id",
            "in": "query",
            "description": "Filter by specific model ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of summaries to return (default: 1000)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of summaries to skip for pagination",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "Filter by specific user ID (admin only)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "System usage summaries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageSummariesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_auth": []
          }
        ]
      }
    },
...
    "/api/usage/records": {
      "get": {
        "tags": [
          "Usage Tracking"
        ],
        "summary": "Get User Usage Records",
        "description": "Retrieve usage records for the authenticated user",
        "operationId": "get_user_usage_records",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to return (default: 50, max: 500)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of records to skip for pagination",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "model",
            "in": "query",
            "description": "Filter by specific model ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "description": "Filter records from this date onwards",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "Filter records up to this date",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "success_only",
            "in": "query",
            "description": "If true, only return successful requests",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Response format (json or csv)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "Filter by specific user ID (admin only)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of usage records",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageRecordsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_auth": []
          },
          {
            "api_key_auth": []
          }
        ]
      }
    },
    "/api/usage/summaries": {
      "get": {
        "tags": [
          "Usage Tracking"
        ],
        "summary": "Get User Usage Summaries",
        "description": "Retrieve pre-computed usage summaries for the authenticated user",
        "operationId": "get_user_usage_summaries",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "description": "Filter summaries from this date onwards",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "Filter summaries up to this date",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "period_type",
            "in": "query",
            "description": "Period type for aggregation (hourly, daily, weekly, monthly)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "model_id",
            "in": "query",
            "description": "Filter by specific model ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of summaries to return (default: 1000)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of summaries to skip for pagination",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "Filter by specific user ID (admin only)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usage summaries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageSummariesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_auth": []
          },
          {
            "api_key_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "description": "API Error response for documentation",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Error code"
          },
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "PeriodType": {
        "type": "string",
        "description": "Period types for usage aggregation",
        "enum": [
          "hourly",
          "daily",
          "weekly",
          "monthly"
        ]
      },
...
      "UsageRecord": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "model_id",
          "endpoint_type",
          "region",
          "request_time",
          "input_tokens",
          "output_tokens",
          "total_tokens",
          "response_time_ms",
          "success"
        ],
        "properties": {
          "cache_read_tokens": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "cache_write_tokens": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "cost_usd": {
            "type": [
              "string",
              "null"
            ]
          },
          "endpoint_type": {
            "type": "string"
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "input_tokens": {
            "type": "integer",
            "format": "int32"
          },
          "model_id": {
            "type": "string"
          },
          "output_tokens": {
            "type": "integer",
            "format": "int32"
          },
          "region": {
            "type": "string"
          },
          "request_time": {
            "type": "string",
            "format": "date-time"
          },
          "response_time_ms": {
            "type": "integer",
            "format": "int32"
          },
          "stop_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "success": {
            "type": "boolean"
          },
          "total_tokens": {
            "type": "integer",
            "format": "int32"
          },
          "user_id": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "UsageRecordsResponse": {
        "type": "object",
        "description": "Response for usage records endpoint",
        "required": [
          "records",
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "limit": {
            "type": "integer",
            "format": "int32",
            "description": "Number of records returned in this page",
            "minimum": 0
          },
          "offset": {
            "type": "integer",
            "format": "int32",
            "description": "Number of records skipped",
            "minimum": 0
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UsageRecord"
            },
            "description": "List of usage records"
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of matching records (for pagination)",
            "minimum": 0
          }
        }
      },
...
      "UsageSummariesResponse": {
        "type": "object",
        "description": "Response for usage summaries endpoint",
        "required": [
          "summaries",
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "limit": {
            "type": "integer",
            "format": "int32",
            "description": "Number of summaries returned in this page",
            "minimum": 0
          },
          "offset": {
            "type": "integer",
            "format": "int32",
            "description": "Number of summaries skipped",
            "minimum": 0
          },
          "summaries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UsageSummary"
            },
            "description": "List of usage summaries"
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of matching summaries (for pagination)",
            "minimum": 0
          }
        }
      },
      "UsageSummary": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "model_id",
          "period_type",
          "period_start",
          "period_end",
          "total_requests",
          "successful_requests",
          "total_input_tokens",
          "total_output_tokens",
          "total_cache_write_tokens",
          "total_cache_read_tokens",
          "total_tokens",
          "avg_response_time_ms",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "avg_response_time_ms": {
            "type": "number",
            "format": "float"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "estimated_cost": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "model_id": {
            "type": "string"
          },
          "period_end": {
            "type": "string",
            "format": "date-time"
          },
          "period_start": {
            "type": "string",
            "format": "date-time"
          },
          "period_type": {
            "$ref": "#/components/schemas/PeriodType"
          },
          "successful_requests": {
            "type": "integer",
            "format": "int32"
          },
          "total_cache_read_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "total_cache_write_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "total_input_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "total_output_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "total_requests": {
            "type": "integer",
            "format": "int32"
          },
          "total_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_id": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key_auth": {
        "type": "http",
        "scheme": "bearer"
      },
      "jwt_auth": {
        "type": "http",
        "scheme": "bearer"
      },
      "x_api_key_auth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}